@@ -1,13 +1,13 @@
 [tool.poetry]
 name = "pyobs-core"
 packages = [{ include = "pyobs" }]
-version = "1.13.0"
+version = "1.17.8"
 description = "robotic telescope software"
 authors = ["Tim-Oliver Husser <thusser@uni-goettingen.de>"]
 license = "MIT"
 
 [tool.poetry.dependencies]
-python = ">=3.9,<3.12"
+python = ">=3.10,<3.13"
 scipy = "^1.11.3"
 astropy = "^5.3.4"
 numpy = "^1.26.0"
@@ -21,12 +21,11 @@ dbus-next = "^0.2.3"
 asyncinotify = "^4.0.2"
 sep = { version = "^1.2.1", platform = "linux", optional = true }
 python-daemon = "^3.0.1"
-reproject = { version = "^0.12.0", optional = true }
 pandas = "^2.1.1"
 astroplan =  "^0.9.1"
 paramiko = "^3.3.1"
 py-expression-eval = "^0.3.14"
-requests = { version = "^2.31.0", optional = true }
+requests = "^2.31.0"
 ccdproc = { version = "^2.4.1", optional = true }
 photutils = { version = "^1.9.0", optional = true }
 lmfit = { version = "^1.2.2", optional = true }
@@ -35,8 +34,8 @@ python-telegram-bot = { version = "^20.6", optional = true }
 aiohttp = "^3.8.6"
 
 [tool.poetry.extras]
-full = ['asyncinotify', 'sep', 'python-daemon', 'reproject', 'pandas', 'paramiko',
-        'requests', 'ccdproc', 'photutils', 'lmfit', 'tornado', 'python-telegram-bot',
+full = ['asyncinotify', 'sep', 'python-daemon', 'pandas', 'paramiko',
+        'ccdproc', 'photutils', 'lmfit', 'tornado', 'python-telegram-bot',
         'aiohttp']
 
 [tool.poetry.dev-dependencies]
