@@ -1,4 +1,4 @@
-import datetime
+from datetime import datetime, timezone
 from typing import List, Optional, Tuple
 
 import numpy as np
@@ -7,16 +7,16 @@ import numpy as np
 class RollingTimeAverage(object):
     def __init__(self, interval: float):
         self._interval = interval
-        self._values: List[Tuple[datetime.datetime, float]] = []
-        self._start_time = datetime.datetime.utcnow()
+        self._values: List[Tuple[datetime, float]] = []
+        self._start_time = datetime.now(timezone.utc)
 
     def clear(self) -> None:
         self._values = []
-        self._start_time = datetime.datetime.utcnow()
+        self._start_time = datetime.now(timezone.utc)
 
     def add(self, value: float) -> None:
         # add value
-        now = datetime.datetime.utcnow()
+        now = datetime.now(timezone.utc)
         self._values.append((now, value))
 
         # clean up
@@ -28,7 +28,7 @@ class RollingTimeAverage(object):
             return None
 
         # get time
-        now = datetime.datetime.utcnow()
+        now = datetime.now(timezone.utc)
 
         # go no values older than now-interval?
         if min_interval:
