@@ -9,9 +9,21 @@ class IMode(Interface, metaclass=ABCMeta):
 
     __module__ = "pyobs.interfaces"
 
+    async def list_mode_groups(self) -> List[str]:
+        """List names of mode groups that can be set. The index is used as the `group` parameter in the individual
+        methods.
+
+        Returns:
+            List of names of mode groups.
+        """
+        return []
+
     @abstractmethod
-    async def list_modes(self, **kwargs: Any) -> List[str]:
+    async def list_modes(self, group: int = 0, **kwargs: Any) -> List[str]:
         """List available modes.
+
+        Args:
+            group: Group number
 
         Returns:
             List of available modes.
@@ -19,11 +31,12 @@ class IMode(Interface, metaclass=ABCMeta):
         ...
 
     @abstractmethod
-    async def set_mode(self, mode: str, **kwargs: Any) -> None:
+    async def set_mode(self, mode: str, group: int = 0, **kwargs: Any) -> None:
         """Set the current mode.
 
         Args:
             mode: Name of mode to set.
+            group: Group number
 
         Raises:
             ValueError: If an invalid mode was given.
@@ -32,8 +45,11 @@ class IMode(Interface, metaclass=ABCMeta):
         ...
 
     @abstractmethod
-    async def get_mode(self, **kwargs: Any) -> str:
+    async def get_mode(self, group: int = 0, **kwargs: Any) -> str:
         """Get currently set mode.
+
+        Args:
+            group: Group number
 
         Returns:
             Name of currently set mode.
