@@ -205,6 +205,11 @@ class AutoFocusSeries(Module, CameraSettingsMixin, IAutoFocus):
         try:
             focus = self._series.fit_focus()
         except Exception as e:
+            # restore initial guess
+            if self._offset:
+                await focuser.set_focus_offset(float(guess))
+            else:
+                await focuser.set_focus(float(guess))
             raise exc.GeneralError(f"Could not calculate best focus: {e}")
 
         # did focus series fail?
@@ -237,7 +242,11 @@ class AutoFocusSeries(Module, CameraSettingsMixin, IAutoFocus):
 
         # take final image?
         if self._final_image:
-            await self._take_image(camera, exposure_time)
+            log.info("Exposing final image at %.2f mm and broadcasting it...", absolute)
+            if isinstance(camera, IExposureTime):
+                await camera.set_exposure_time(exposure_time)
+            if isinstance(camera, IData):
+                return await camera.grab_data()
 
         # return result
         return focus[0], focus[1]
