@@ -1,8 +1,11 @@
 from .script import Script
-from .skyflats import SkyFlats
-from .darkbias import DarkBias
-from .sequential import SequentialRunner
-from .parallel import ParallelRunner
-from .selector import SelectorScript
-from .conditional import ConditionalRunner
+from .cases import CasesRunner
 from .callmodule import CallModule
+from .conditional import ConditionalRunner
+from .darkbias import DarkBias
+from .debugtrigger import DebugTriggerRunner
+from .parallel import ParallelRunner
+from .log import LogRunner
+from .selector import SelectorScript
+from .sequential import SequentialRunner
+from .skyflats import SkyFlats
