@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from datetime import datetime, timezone
 import logging
-from typing import Any, Dict, Optional, List, TYPE_CHECKING
+from typing import Any, Dict, Optional, List, TYPE_CHECKING, Union
 
 if TYPE_CHECKING:
     from pyobs.robotic import TaskRunner, TaskSchedule, TaskArchive
@@ -11,40 +11,38 @@ from pyobs.robotic.scripts import Script
 log = logging.getLogger(__name__)
 
 
-class ConditionalRunner(Script):
-    """Script for running an if condition."""
+class CasesRunner(Script):
+    """Script for distinguishing cases."""
 
     __module__ = "pyobs.modules.robotic"
 
     def __init__(
         self,
-        condition: str,
-        true: Dict[str, Any],
-        false: Optional[Dict[str, Any]] = None,
+        expression: str,
+        cases: Dict[Union[str, int, float], Any],
         **kwargs: Any,
     ):
-        """Initialize a new ConditionalRunner.
+        """Initialize a new CasesRunner.
 
         Args:
-            condition: condition to check
-            true: script to run if condition is evaluated as True
-            false: script to run otherwise.
+            expression: expression to check
+            cases: dictionary with cases
         """
         Script.__init__(self, **kwargs)
-        self.condition = condition
-        self.true = true
-        self.false = false
+        self.expression = expression
+        self.cases = cases
 
     def __get_script(self) -> Script:
         # evaluate condition
-        ret = eval(self.condition, {"now": datetime.now(timezone.utc)})
+        value = eval(self.expression, {"now": datetime.now(timezone.utc), "config": self.configuration})
 
-        # run scripts
-        if ret:
-            return self.get_object(self.true, Script)
+        # check in cases
+        if value in self.cases:
+            return self.get_object(self.cases[value], Script, configuration=self.configuration)
+        elif "else" in self.cases:
+            return self.get_object(self.cases["else"], Script, configuration=self.configuration)
         else:
-            if self.false is not None:
-                return self.get_object(self.false, Script)
+            raise ValueError("Invalid choice")
 
     async def can_run(self) -> bool:
         script = self.__get_script()
@@ -60,4 +58,4 @@ class ConditionalRunner(Script):
         await script.run(task_runner, task_schedule, task_archive)
 
 
-__all__ = ["ConditionalRunner"]
+__all__ = ["CasesRunner"]
